Refactors AuthStateInMemoryStore into a mock package
